css/css-values/tree-counting/calc-sibling-function-in-shadow-dom.html

<!DOCTYPE html>
<html>
<head>
  <title>CSS Values Test: sibling-index() and sibling-count() in flat tree</title>
  <link rel="help" href="https://drafts.csswg.org/css-values-5/#tree-counting" />
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
  <style>
    #host > * {
      grid-column-start: sibling-index();
      grid-row-start: sibling-count();
    }
  </style>
  <div id="host">
    <div>Some element.</div>
    <div>Some other element.</div>
    <div slot="unknown">This should be ignored.</div>
    <div id="target">Some third element.</div>
  </div>
  <script>
    const shadowRoot = host.attachShadow({mode: 'open'});
    shadowRoot.innerHTML = `
      <style>
        slot::slotted(*), #shadow-root-child {
          z-index: sibling-index();
          order: sibling-count();
        }
      </style>
      <div id="shadow-root-child">
        <div>Some text before the slot.</div>
        <slot></slot>
      </div>
    `;

    test(() => {
      assert_equals(getComputedStyle(target).zIndex, '4');
      assert_equals(getComputedStyle(target).order, '4');
      assert_equals(getComputedStyle(target).gridColumnStart, '4');
      assert_equals(getComputedStyle(target).gridRowStart, '4');
    }, 'Host children have sibling-index() and sibling-count() based on the DOM tree order');

    test(() => {
      const shadowRootChild = shadowRoot.getElementById('shadow-root-child');
      assert_equals(getComputedStyle(shadowRootChild).zIndex, '2', 'sibling-index() for direct shadow root child');
      assert_equals(getComputedStyle(shadowRootChild).order, '2', 'sibling-count() for direct shadow root child');
    }, 'Direct shadow root children have correct sibling-index() and sibling-count()')
</script>
